requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
pyjwt==2.10.1
argon2-cffi==23.1.0
pydantic[email]==2.10.3
python-dotenv==1.0.1
jinja2==3.1.4
pytest==8.3.4
httpx==0.27.2
